Minor improvement
On Synology NAS, python 3.8 numpy did not like the fact that linspace had not an int for the grid number of points. Now this is enforced

+++ model.py
@@ -179,7 +179,7 @@ def test_model_core_alt():
 	fmin=50
 	fmax=400
 	resol=0.05
-	freq=np.linspace(fmin, fmax, (fmax-fmin)/resol +1)
+	freq=np.linspace(fmin, fmax, int((fmax-fmin)/resol +1))
 	
 	inc=45
